+++ game/ato/flightplans/waypointbuilder.py
@@ -14,6 +14,7 @@ from typing import (
 
 from dcs.mapping import Point, Vector2
 
+from game.ato.flightplans._common_ctld import generate_random_ctld_point
 from game.ato.flightwaypoint import AltitudeReference, FlightWaypoint
 from game.ato.flightwaypointtype import FlightWaypointType
 from game.theater import (
@@ -23,6 +24,7 @@ from game.theater import (
     TheaterGroundObject,
     TheaterUnit,
 )
+from game.theater.interfaces.CTLD import CTLD
 from game.utils import Distance, meters, nautical_miles, feet
 
 if TYPE_CHECKING:
@@ -557,10 +559,14 @@ class WaypointBuilder:
         """Creates a cargo stop waypoint.
         This waypoint is used by AirLift as a landing and stopover waypoint
         """
+        if isinstance(control_point, CTLD) and control_point.ctld_zones:
+            pos = generate_random_ctld_point(control_point)
+        else:
+            pos = control_point.position
         return FlightWaypoint(
             "CARGOSTOP",
             FlightWaypointType.CARGO_STOP,
-            control_point.position,
+            pos,
             meters(0),
             "RADIO",
             description=f"Stop for cargo at {control_point.name}",