+++ game/theater/interfaces/CTLD.py
@@ -1,0 +1,7 @@
+from typing import Optional, List, Tuple
+
+from dcs import Point
+
+
+class CTLD:
+    ctld_zones: Optional[List[Tuple[Point, float]]] = None